mini_claw/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mini_claw"
version = "0.1.0"
description = "Codeclaw cloud platform - remote AI automation powered by mini_claude"
readme = "docs/CODECLAW_PLAN.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["ai", "automation", "telegram", "cloud", "agent"]

dependencies = [
    "python-telegram-bot>=20.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[project.scripts]
mini_claw = "script.main:cli"
mclaw = "script.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["script*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "W", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"
